+++ backend/common/enums/api_settings_names.py
@@ -5,6 +5,7 @@ class ApiSettingsNamesEnum(Enum):
     api_version = 'app_version'
     db_version = 'db_version'
     webapp_main_color = 'webapp_main_color'
+    index_page = 'index_page'
 
 
 API_SETTINGS_NAMES_LIST: Final = [x.value for x in ApiSettingsNamesEnum]