+++ backend/cwa_lib/pydantic_schemas/su_manage_api_settings.py
@@ -1,5 +1,6 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, field_validator
 from .generic_table import TableQueryResult
+from cwa_lib.validators.api_settings import validate__name
 
 
 class SuManageApiSettingsRead(BaseModel):
@@ -11,7 +12,19 @@ class SuManageApiSettingsRead(BaseModel):
 
 SuManageApiSettingsQueryResult = TableQueryResult[SuManageApiSettingsRead]
 
+class SuManageApiSettingsCreate(BaseModel):
+    name: str
+    value: str
+    @field_validator("name")
+    @classmethod
+    def validate__name(cls, v: str) -> str:
+        return validate__name(v)
+
 
 class SuManageApiSettingsUpdate(BaseModel):
     name: str
     value: str
+    @field_validator("name")
+    @classmethod
+    def validate__name(cls, v: str) -> str:
+        return validate__name(v)